slightly altered the order of how we pre-process text. Now the order is: 1. Lower-casing 2. Tokenize 3. Stopword Removal 4. Punctuation Removal 5. Lemmatize
There's still a minor bug if you input lots of white spaces, shouldn't be too difficult to fix.
